refactor: Add version 1.5.0 update notification
This commit adds a new function `versionOneFiveZeroUpdateNotification` to the `utils.ts` file. This function displays a notification to the user when the extension is updated to version 1.5.0. The notification includes information about the changes in this version and recommends signing out and signing back in to ensure all features work correctly. The function checks the current version of the extension and a flag in the local storage to determine if the notification should be displayed. If the flag is not set and the current version is 1.5.0, the notification is shown and the flag is set to prevent showing the notification again.