Feat: Add course adding "Loading" text and progress bar  (#151)
* feat: Show progress of course add

This commit adds code to update the `courseAddProgress` value in the `chrome.storage.local` object at various points in the `nameSearchApi.ts` and `idSearchApi.ts` files. This allows for tracking the progress of adding a course to the scheduler.

* feat: Switch sending through storage to custom events

* lint and prettier

* fix: handleCourseLoading function not called in extractSection when details not found.

The handleCourseLoading function was not being called in the extractSection function of the utils.ts file. This caused the course loading state to not be updated correctly. This commit fixes the issue by adding the missing handleCourseLoading function call.